+++ anansescanpy/__init__.py
@@ -1,11 +1,11 @@
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
-from .anansescanpy import export_CPM_scANANSE
-from .anansescanpy import export_ATAC_scANANSE
-from .anansescanpy import config_scANANSE
-from .anansescanpy import DEGS_scANANSE
-from .anansescanpy import per_cluster_df
-from .anansescanpy import import_scanpy_scANANSE
-from .anansescanpy import Maelstrom_Motif2TF
-from .anansescanpy import import_scanpy_maelstrom
-from .anansescanpy import Factor_Motif_Plot
+from .anansescanpy_export import export_CPM_scANANSE
+from .anansescanpy_export import export_ATAC_scANANSE
+from .anansescanpy_export import config_scANANSE
+from .anansescanpy_export import DEGS_scANANSE
+from .anansescanpy_import import per_cluster_df
+from .anansescanpy_import import import_scanpy_scANANSE
+from .anansescanpy_import import import_scanpy_maelstrom
+from .anansescanpy_motif import Maelstrom_Motif2TF
+from .anansescanpy_motif import Factor_Motif_Plot
