+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "anansescanpy"
-version = "0.2.4"
+version = "0.2.5"
 description = "implementation of scANANSE for scanpy objects in Python"
 authors = ["J Arts (Arts-of-coding) <julian.armando.arts@gmail.com>"]
 license = "apache 2.0"
@@ -13,11 +13,11 @@ keywords = ["GRN","single-cell","gene expression","open chromatin","arts-of-codi
 python = "^3.8"
 scanpy = "^1.9.1"
 #numba = "^0.56.3"  # numpy
-numpy = "^1.23.3"
+numpy = "< 1.24"
 anndata = "^0.8.0"  # for convenience
 scipy = "^1.9.1"
 pandas = "^1.4.4"
-jupyterlab = "^3.3.4"  # for convenience
+jupyterlab = "^3.4.6"  # for convenience
 
 [[tool.poetry.source]]
 name = "anansescanpy"
