+++ requirements.yaml
@@ -7,11 +7,13 @@ dependencies:
   - python >=3.8
   - scanpy >=1.9.1
 #  - numba >=0.56.3  # numpy
-  - numpy >=1.23.3
+  - numpy <1.24
   - anndata >=0.8.0  # for convenience
   - scipy >=1.9.1
   - pandas >=1.4.4
-  - jupyterlab >=3.3.4  # for convenience
+  - jupyterlab >=3.4.6  # for convenience
+  - packaging >=21.3
+  - jinja2 >=3.0
 
 # for development
   - black
