+++ tests/test_anansescanpy.py
@@ -2,4 +2,4 @@ from anansescanpy import __version__
 
 
 def test_version():
-    assert __version__ == '0.2.4'
+    assert __version__ == '0.2.5'
